Add files via upload
This compilation of the Philippines' national statements at the United Nations General Assembly from 1946 to 2024 offers a unique longitudinal perspective on Philippine foreign policy, diplomatic advocacies and multilateral priorities over nearly eight decades since the founding of the United Nations in 1945.


Spanning successive presidential administrations and seismic global shifts—from the post-World War II reconstruction era to the challenges of the threat of war and conflict, decolonization, debt and economic inequality to contemporary crises like climate change and pandemics —these policy statements capture the nation’s aspirations and advocacies on the global stage. 


Several thematic constants emerge: a steadfast advocacy for peace, decolonization, and the rule of law; a consistent call for nuclear disarmament and equitable development; and a growing emphasis on human rights, migrant protection, and environmental stewardship. Across time, the Philippines has risen to respond to the exigencies of each era: invoking solidarity during the Cold War, asserting sovereignty in periods of regional instability, embracing cooperation in the face of global crises and pursuing the development of international law as anchor of peace, justice and equity.